refactor: isolate and reuse CalculateCartTotal function in payment and cart page
